TS Panel_error and dashboard_constant  (#22141)
* Panel migration to ts

Added the getKibanaVersion to index.d.ts file
Panel error now in ts

* test

* Panel migration to ts

Added the getKibanaVersion to index.d.ts file
Panel error now in ts

* deleted dashboard_constants.js

* removed comment in dashboard_constants
