+++ src/core_plugins/kibana/public/dashboard/dashboard_constants.ts
@@ -17,22 +17,17 @@
  * under the License.
  */
 
-import React from 'react';
-import PropTypes from 'prop-types';
-
-export function PanelError({ error }) {
-  return (
-    <div className="dshPanel__error panel-content">
-      <span aria-hidden="true" className="kuiIcon fa-exclamation-triangle"/>
-      <span>{error}</span>
-    </div>
-  );
-}
-
-PanelError.propTypes = {
-  error: PropTypes.oneOfType([
-    PropTypes.string,
-    PropTypes.node
-  ]),
+export const DashboardConstants = {
+  ADD_VISUALIZATION_TO_DASHBOARD_MODE_PARAM: 'addToDashboard',
+  NEW_VISUALIZATION_ID_PARAM: 'addVisualization',
+  LANDING_PAGE_PATH: '/dashboards',
+  CREATE_NEW_DASHBOARD_URL: '/dashboard',
 };
+export const DASHBOARD_GRID_COLUMN_COUNT = 48;
+export const DASHBOARD_GRID_HEIGHT = 20;
+export const DEFAULT_PANEL_WIDTH = DASHBOARD_GRID_COLUMN_COUNT / 2;
+export const DEFAULT_PANEL_HEIGHT = 15;
 
+export function createDashboardEditUrl(id: string) {
+  return `/dashboard/${id}`;
+}
