+++ src/core_plugins/kibana/public/dashboard/panel/panel_error.tsx
@@ -17,17 +17,17 @@
  * under the License.
  */
 
-export const DashboardConstants = {
-  ADD_VISUALIZATION_TO_DASHBOARD_MODE_PARAM: 'addToDashboard',
-  NEW_VISUALIZATION_ID_PARAM: 'addVisualization',
-  LANDING_PAGE_PATH: '/dashboards',
-  CREATE_NEW_DASHBOARD_URL: '/dashboard',
-};
-export const DASHBOARD_GRID_COLUMN_COUNT = 48;
-export const DASHBOARD_GRID_HEIGHT = 20;
-export const DEFAULT_PANEL_WIDTH = DASHBOARD_GRID_COLUMN_COUNT / 2;
-export const DEFAULT_PANEL_HEIGHT = 15;
+import React from 'react';
 
-export function createDashboardEditUrl(id) {
-  return `/dashboard/${id}`;
+export interface PanelErrorProps {
+  error: string | React.ReactNode;
+}
+
+export function PanelError({ error }: PanelErrorProps) {
+  return (
+    <div className="dshPanel__error panel-content">
+      <span aria-hidden="true" className="kuiIcon fa-exclamation-triangle" />
+      <span>{error}</span>
+    </div>
+  );
 }
